Client Tests Fully Done.

+++ server/services/client.service.js
@@ -47,22 +47,27 @@ export const ClientService = {
       );
     }
 
-    await ClientRepo.findByEmail(data.email).OrElseThrow(
-      new Error("Client with this email already exists"),
-    );
+    const existingClient = await ClientRepo.findByEmail(data.email);
 
-    const clientUser = await UserRepo.findById(data.user.id).OrElseThrow(
-      new Error("User not found"),
-    );
+    if (existingClient)
+      throw new Error("Client with this email already exists");
 
-    if (clientUser.role != "CLIENT_ADMIN")
-      throw new Error(
-        "Current role does not have the permission to become admin for this client.",
-      );
+    if (data.assignedAdmin != null) {
+      const clientUser = await UserRepo.findById(data.user?.id);
 
-    await ClientRepo.findByUser(clientUser.id).OrElseThrow(
-      new Error("User is already assigned to a client"),
-    );
+      if (clientUser) {
+        if (clientUser.role != "CLIENT_ADMIN")
+          throw new Error(
+            "Current role does not have the permission to become admin for this client.",
+          );
+
+        const user = await ClientRepo.findByUser(clientUser.id);
+
+        if (user) throw new Error("User is already assigned to a client.");
+      } else {
+        throw new Error("User not found");
+      }
+    }
 
     const client = await ClientRepo.create({
       ...data,
@@ -78,7 +83,7 @@ export const ClientService = {
       return await ClientRepo.update(data.id, { set: { deleteRequest: true } });
     } else {
       throw new Error(
-        "Current role does not have the permission to add Clients",
+        "Current role does not have the permission to request client deletion.",
       );
     }
   },
@@ -88,31 +93,47 @@ export const ClientService = {
     // SUPER_ADMIN can access everything
     if (user.role !== "SUPER_ADMIN") {
       throw new Error(
-        "Current role does not have the permission to add Clients",
+        "Current role does not have the permission to delete Clients.",
       );
     }
 
-    await ClientRepo.findById(data.id).OrElseThrow(
-      new Error("Client not found"),
-    );
+    const client = await ClientRepo.findById(data.id);
+    if (!client) throw new Error("Client not found");
+
+    if (!client.deleteRequest)
+      throw new Error("Delete request not found for this client.");
+
+    return await ClientRepo.delete(data.id);
+  },
+
+  async forceDeleteClientBySuperAdmin(data, context) {
+    const { user } = context;
+
+    if (user.role !== "SUPER_ADMIN") {
+      throw new Error(
+        "Current role does not have the permission to delete Clients.",
+      );
+    }
+
+    const client = await ClientRepo.findById(data.id);
+    if (!client) throw new Error("Client not found");
 
     return await ClientRepo.delete(data.id);
   },
   async updateClient(data, context) {
     const { user } = context;
 
-    // SUPER_ADMIN can access everything
     if (user.role == "USER") {
       throw new Error(
         "Current role does not have the permission to update Clients",
       );
     }
 
-    const client = await ClientRepo.findById(data.id).OrElseThrow(
-      "Client not found",
-    );
+    const client = await ClientRepo.findById(data.id);
 
-    if (user.id == client.assignedUser.id || user.role == "SUPER_ADMIN") {
+    if (!client) throw new Error("Client not found");
+
+    if (user.id == client.assignedAdmin || user.role == "SUPER_ADMIN") {
       return await ClientRepo.update(data.id, data);
     }
